🔨 [functional] Move `session.poetry` tests to separate module
