+++ tests/functional/test_session.py
@@ -129,18 +129,6 @@ def test_path_dependency(shared_datadir: Path) -> None:
     assert set(expected) == set(packages)
 
 
-def test_passthrough_env(project: Project) -> None:
-    """It stores requirements even without a virtualenv directory."""
-    # https://github.com/cjolowicz/nox-poetry/issues/347
-
-    @nox_poetry.session(venv_backend="none")
-    def test(session: nox_poetry.Session) -> None:
-        """Export the requirements."""
-        session.poetry.export_requirements()
-
-    run_nox_with_noxfile(project, [test], [nox_poetry])
-
-
 def test_no_install(project: Project) -> None:
     """It skips installation when --no-install is passed."""
 
@@ -148,23 +136,6 @@ def test_no_install(project: Project) -> None:
     def test(session: nox_poetry.Session) -> None:
         """Install the local package."""
         session.install(".")
-
-    run_nox_with_noxfile(project, [test], [nox_poetry])
-    run_nox_with_noxfile(project, [test], [nox_poetry], "-R")
-
-    expected = [project.package, *project.dependencies]
-    packages = list_packages(project, test)
-
-    assert set(expected) == set(packages)
-
-
-def test_no_install_installroot(project: Project) -> None:
-    """It skips installation when --no-install is passed."""
-
-    @nox_poetry.session
-    def test(session: nox_poetry.Session) -> None:
-        """Install the local package."""
-        session.poetry.installroot()
 
     run_nox_with_noxfile(project, [test], [nox_poetry])
     run_nox_with_noxfile(project, [test], [nox_poetry], "-R")
